refactor

+++ basic-network/generate.sh
@@ -4,55 +4,50 @@
 #
 # SPDX-License-Identifier: Apache-2.0
 #
-export PATH=$GOPATH/src/github.com/hyperledger/fabric/build/bin:${PWD}/../bin:${PWD}:$PATH
+export PATH=$GOPATH/src/github.com/hyperledger/fabric/build/bin:~/bin:$PATH
 export FABRIC_CFG_PATH=${PWD}
 CHANNEL_NAME=mychannel
 
-# remember sk, by Jack
-sk_old=$(find crypto-config/peerOrganizations/org1.example.com/ca -name "*_sk" -exec basename {} \;)
+while getopts n opt
+do
+    case $opt in
+        n)  RENEW=1;;
+        ?)  ;;
+    esac
+done
 
-# remove previous crypto material and config transactions
-rm -fr config/*
-rm -fr crypto-config/*
+if [[ x$RENEW = x1 ]]
+then
+    echo Re-generate crypto-marterial...
 
-# generate crypto material
-cryptogen generate --config=./crypto-config.yaml
-if [ "$?" -ne 0 ]; then
-  echo "Failed to generate crypto material..."
-  exit 1
-fi
+    # remember sk, by Jack
+    sk_old=$(find crypto-config/peerOrganizations/org1.example.com/ca -name "*_sk" -exec basename {} \;)
 
-# update sk, by Jack
-sk_new=$(find crypto-config/peerOrganizations/org1.example.com/ca -name "*_sk" -exec basename {} \;)
-sed -i "s/"$sk_old"/"$sk_new"/" ./docker-compose.yml
+    # remove previous crypto material and config transactions
+    rm -fr config/*
+    rm -fr crypto-config/*
 
-# generate genesis block for orderer
-configtxgen -profile OneOrgOrdererGenesis -outputBlock ./config/genesis.block
-if [ "$?" -ne 0 ]; then
-  echo "Failed to generate orderer genesis block..."
-  exit 1
-fi
-
-gen_crypto_4_channel () {
-    CHANNEL_NAME=$1
-
-    # generate channel configuration transaction
-    configtxgen -profile OneOrgChannel -outputCreateChannelTx ./config/channel_$CHANNEL_NAME.tx -channelID $CHANNEL_NAME
+    # generate crypto material
+    cryptogen generate --config=./crypto-config.yaml
     if [ "$?" -ne 0 ]; then
-        echo "Failed to generate channel configuration transaction..."
+        echo "Failed to generate crypto material..."
         exit 1
     fi
 
-    # generate anchor peer transaction
-    configtxgen -profile OneOrgChannel -outputAnchorPeersUpdate ./config/Org1MSPanchors_$CHANNEL_NAME.tx -channelID $CHANNEL_NAME -asOrg Org1MSP
+    # update sk, by Jack
+    sk_new=$(find crypto-config/peerOrganizations/org1.example.com/ca -name "*_sk" -exec basename {} \;)
+    sed -i "s/"$sk_old"/"$sk_new"/" ./docker-compose.yml
+
+    # generate genesis block for orderer
+    configtxgen -profile OneOrgOrdererGenesis -outputBlock ./config/genesis.block
     if [ "$?" -ne 0 ]; then
-        echo "Failed to generate anchor peer update for Org1MSP..."
+        echo "Failed to generate orderer genesis block..."
         exit 1
     fi
-}
+fi
 
 . ./fabric.conf
-for chan in $CHANNEL
+for chan in `for i in ${CHANNEL[@]}; do echo $i; done | sort -u`
 do
     gen_crypto_4_channel $chan
 done
